requirements.txt
opencv-python
numpy
tk
scikit-learn

requests
replicate
pillow
openai